Fix map activity tests
